order calculation modified
The price of new order is calculated relative to price of the last
trade, instead of price of equilibrium. This better handles spread - now
the spread is not affected by the sliding_pos. Shifted equilibrium now
appears as adjusted size of the order. If the size is calculated as
negative, the order is not placed.